[feat] 첫 번쨰 피드백 반영
